Plugin is now independent from GWT version
As a result, gwt-dev and gwt-codeserver must now be added as project
dependencies.

Plugin now also requires Java 7 (next version of GWT will require it too).

Travis CI builds with the last two releases and the latest snapshots of
GWT (2.6.1, 2.7.0 and 2.8.0-SNAPSHOT), in all Java environments (openjdk7,
oraclejdk7 and oraclejdk8).