Back to using runtime scope for (super)devmode, but make it configurable.
Using different scopes for compiler vs. devmode makes it possible to have
slightly different dependencies in those cases. In most cases, dependencies
are expected to be in the compile scope so the classpaths would be
identical between compiler and devmode.
For "legacy" projects that mix client and server code in the same Maven
module, making it configurable allows for having GWT dependencies in the
provided scope. This will help in migrating those projects to this plugin
one step at a time.