Remove devMode from gwt:test as it's now deprecated in 2.8
People using 2.7 can turn it back on with an explicit 'testArgs':

    <testArgs>
      <arg>-devMode</arg>
    </testArgs>

and can get back the property-triggered behavior with:

    <testArgs>
      <arg>${gwt.test.devMode}</arg>
    </testArgs>

(setting the property to either -devMode or -nodevMode –or -prod or -web–
to turn it on/off from the command line). There are of course alternatives
using Maven profiles for the adventurous.