Fix gwt-version-specific integration tests
For some reason, Maven only reads the gwtVersion from the POM when
filtering POMs for integration tests, so all three Travis builds were
actually building against GWT 2.7.0. I swear it worked with some version
of Maven (3.0 ?) but it fails with 3.2.5 (from Travis) and 3.3.9 (local).

Note: this also applies to the site, but we don't care as we always build
the site with the property value from the POM.