Fix annotation-processing in IT tests (specifically gwt-lib)
When source/target versions are less than 6 in maven-compiler-plugin,
irrespective of the JavaC version, the '-s' argument won't be passed
to JavaC, so generated source files are written into target/classes,
generating a false-positive in the postbuild.bsh as those sources are
then packaged into the gwt-lib JAR whereas gwt:import-sources is run
too early.

This commit therefore breaks the build.