Use referenced project's artifact for gwt-lib in devmode/codeserver
In a reactor build, Maven will resolve <type>jar</type> dependencies
(without classifier) to the referenced project's build output
directory (target/classes); but it will resolve everything else
from the local or remote repositories. When a gwt-lib is published
as SNAPSHOT, it will thus be used instead of the referenced project,
meaning that you'd end up with stale code and the edit/refresh
developer experience is broken.

As a workaround, dependencies with type gwt-lib and a corresponding
project reference are replaced with the referenced projects build
output directory in the devmode/codeserver classpath.