Add support for Maven Toolchains to select the JVM for forked processes
This bumps the Maven requirement to version 3.3.1, which
was released more than 4 years ago already.

There's a small discrepancy between gwt:compile, gwt:devmode,
and gwt:codeserver on one hand, and gwt:test on the other hand,
in that gwt:test relies on Surefire so it doesn't have a
jdkToolchain to configure an plugin execution scoped toolchain.

Fixes #130