Add tests for different maven versions
This checks the latest bugfix release of each minor release of maven
since the minimum supported (presently 3.3.1). The CI job is based on
test-gwt, with substitutions for the maven version (using mvnw) instead
of replacing the GWT version.

Also updated the JDK builds to check 18 (current latest), and updated
the GWT builds to test 2.9.0.